+++ smartorm-spring-boot-starter/src/test/java/io/github/forgottenlab/smartorm/autoconfigure/defaultscan/DefaultScanApplication.java
@@ -1,0 +1,10 @@
+package io.github.forgottenlab.smartorm.autoconfigure.defaultscan;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+
+@SpringBootConfiguration(proxyBeanMethods = false)
+@EnableAutoConfiguration(exclude = DataSourceAutoConfiguration.class)
+public class DefaultScanApplication {
+}